+++ docs/source/conf.py
@@ -18,12 +18,8 @@
 # -- Project information -----------------------------------------------------
 
 project = 'attributecode'
-copyright = '2020 nexB Inc.'
-author = 'nexB Inc.'
-
-version = '0.0.1'
-# The full version, including alpha/beta/rc tags
-release = version
+copyright = 'nexb Inc.'
+author = 'nexb Inc.'
 
 
 # -- General configuration ---------------------------------------------------
@@ -31,7 +27,8 @@ release = version
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = []
+extensions = [
+]
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
@@ -42,15 +39,12 @@ templates_path = ['_templates']
 exclude_patterns = []
 
 
-master_doc = 'index'
-
 # -- Options for HTML output -------------------------------------------------
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-# html_theme = 'alabaster'
-html_theme = "sphinx_rtd_theme"
+html_theme = 'sphinx_rtd_theme'
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
@@ -63,7 +57,7 @@ html_context = {
         ],
     "display_github": True,
     "github_user": "nexB",
-    "github_repo": "attributecode",
+    "github_repo": "nexb-skeleton",
     "github_version": "develop",  # branch
     "conf_py_path": "/docs/source/",  # path in the checkout to the docs root
     }