+++ CHANGELOG.rst
@@ -1,5 +1,24 @@
 Release notes
 -------------
-### Version 0.0.0
 
-*xxxx-xx-xx* -- Initial release.
+### Version 2.1.1
+
+2021-06-23
+ - Update with the latest skeleton which incude the configuration scripts and others
+ - Remove the thirdparty/
+
+### Version 2.0.1
+
+2021-05-28
+ - Update template wording
+
+### Version 2.0.0
+
+2021-05-28
+ - Pass a list of dictionaries when doing rendering instead of using field objects
+ - Output number of component generated in the attribution if default_html.template is used.
+ 
+
+### Version 1.0.0
+
+2021-03-31 -- Initial release.